From QMap<QRgb, int> to QMap<QRgb, long>

+++ Image-Recovery/Image-Recovery/src/algorithm/datamanager.cpp
@@ -60,11 +60,11 @@ void DataManager::setReference(const QVector<QRgb> &reference) {
     DataManager::reference = reference;
 }
 
-const QMap<QRgb, int> &DataManager::getColorTableReference() const {
+const QMap<QRgb, long> &DataManager::getColorTableReference() const {
     return colorTableReference;
 }
 
-void DataManager::setColorTableReference(const QMap<QRgb, int> &colorTableReference) {
+void DataManager::setColorTableReference(const QMap<QRgb, long> &colorTableReference) {
     DataManager::colorTableReference = colorTableReference;
 }
